Move StreamReader factories
