+++ ingestor/ingestor-default/src/main/scala/za/co/absa/hyperdrive/ingestor/implementation/reader/kafka/KafkaStreamReader.scala
@@ -17,11 +17,14 @@
 
 package za.co.absa.hyperdrive.ingestor.implementation.reader.kafka
 
+import org.apache.commons.configuration2.Configuration
 import org.apache.commons.lang3.StringUtils
 import org.apache.logging.log4j.LogManager
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.streaming.DataStreamReader
-import za.co.absa.hyperdrive.ingestor.api.reader.StreamReader
+import za.co.absa.hyperdrive.ingestor.api.reader.{StreamReader, StreamReaderFactory}
+import za.co.absa.hyperdrive.shared.configurations.ConfigurationsKeys.KafkaStreamReaderKeys.{KEY_BROKERS, KEY_KEYSTORE_LOCATION, KEY_KEYSTORE_PASSWORD, KEY_KEY_PASSWORD, KEY_SECURITY_PROTOCOL, KEY_TOPIC, KEY_TRUSTSTORE_LOCATION, KEY_TRUSTSTORE_PASSWORD, rootComponentConfKey, rootFactoryOptionalConfKey, rootFactoryOptionalKafkaKey}
+import za.co.absa.hyperdrive.shared.utils.ConfigUtils.{getOrNone, getOrThrow, getSeqOrThrow}
 
 private[reader] object KafkaStreamReaderProps {
   val STREAM_FORMAT_KAFKA_NAME  = "kafka"
@@ -37,7 +40,7 @@ private[reader] object KafkaStreamReaderProps {
   * @param brokers String containing the brokers
   * @param extraConfs Extra configurations, e.g. SSL params.
   */
-class KafkaStreamReader(val topic: String, val brokers: String, val extraConfs: Map[String,String]) extends StreamReader {
+private[reader] class KafkaStreamReader(val topic: String, val brokers: String, val extraConfs: Map[String,String]) extends StreamReader {
 
   private val logger = LogManager.getLogger()
 
@@ -83,4 +86,90 @@ class KafkaStreamReader(val topic: String, val brokers: String, val extraConfs: 
   }
 
   override def getSourceName: String = s"Kafka topic: $topic"
+}
+
+object KafkaStreamReader extends StreamReaderFactory {
+  private val logger = LogManager.getLogger
+
+  override def apply(conf: Configuration): StreamReader = {
+    val topic = getTopic(conf)
+    val brokers = getBrokers(conf)
+    val extraOptions = getExtraOptions(conf)
+
+    logger.info(s"Going to create KafkaStreamReader with: topic='$topic', brokers='$brokers', extraOptions=${filterKeysContaining(extraOptions, exclusionToken = "password")}")
+
+    new KafkaStreamReader(topic, brokers, extraOptions)
+  }
+
+  private def getTopic(configuration: Configuration): String = {
+    getOrThrow(KEY_TOPIC, configuration, errorMessage = s"Topic not found. Is $KEY_TOPIC defined?")
+  }
+
+  private def getBrokers(configuration: Configuration): String = {
+    val brokers = getSeqOrThrow(KEY_BROKERS, configuration, errorMessage = s"Brokers not found. Is $KEY_BROKERS defined?")
+    brokers.mkString(",")
+  }
+
+  private def getPredefinedExtraOptions(configuration: Configuration): Map[String,String] = {
+    val securityKeys = Seq(KEY_SECURITY_PROTOCOL, KEY_TRUSTSTORE_LOCATION, KEY_TRUSTSTORE_PASSWORD, KEY_KEYSTORE_LOCATION, KEY_KEYSTORE_PASSWORD, KEY_KEY_PASSWORD)
+
+    val extraConfs = securityKeys.foldLeft(Map[String,String]()) {
+      case (map,securityKey) =>
+        getOrNone(securityKey, configuration) match {
+          case Some(value) => map + (tweakKeyName(securityKey) -> value)
+          case None => map
+        }
+    }
+
+    if (extraConfs.isEmpty || extraConfs.size == securityKeys.size) {
+      extraConfs
+    }
+    else {
+      logger.warn(s"Assuming no security settings, since some appear to be missing: {${findMissingKeys(securityKeys, extraConfs)}}")
+      Map[String,String]()
+    }
+  }
+
+  private def getExtraOptions(configuration: Configuration): Map[String,String] = {
+    val optionalKeys = getKeysFromPrefix(configuration, rootFactoryOptionalKafkaKey)
+
+    val extraConfs = optionalKeys.foldLeft(Map[String,String]()) {
+      case (map,securityKey) =>
+        getOrNone(securityKey, configuration) match {
+          case Some(value) => map + (tweakOptionKeyName(securityKey) -> value)
+          case None => map
+        }
+    }
+
+    if (extraConfs.isEmpty || extraConfs.size == optionalKeys.size) {
+      extraConfs
+    }
+    else {
+      logger.warn(s"Assuming no security settings, since some appear to be missing: {${findMissingKeys(optionalKeys, extraConfs)}}")
+      Map[String,String]()
+    }
+  }
+
+  private def getKeysFromPrefix(configuration: Configuration, prefix: String): Seq[String] = {
+    val optionalKeys = configuration.getKeys(prefix)
+
+    if (optionalKeys != null) {
+      import scala.collection.JavaConverters._
+      optionalKeys.asScala.toSeq
+    } else {
+      Seq[String]()
+    }
+  }
+
+  private def findMissingKeys(keys: Seq[String], map: Map[String,String]): Seq[String] = keys.filterNot(map.contains(_))
+
+  private def tweakKeyName(key: String): String = {
+    key.replace(s"$rootComponentConfKey.", "") // remove the component root configuration key
+  }
+
+  private def tweakOptionKeyName(key: String): String = {
+    key.replace(s"$rootFactoryOptionalConfKey.", "") // remove the component root.option configuration key
+  }
+
+  private def filterKeysContaining(map: Map[String,String], exclusionToken: String): Map[String,String] = map.filterKeys(!_.contains(exclusionToken))
 }
